Refactor import paths for improved consistency
