+++ packages/core-ui/src/components/ui/alert-dialog/alert-dialog.tsx
@@ -1,8 +1,8 @@
 import * as AlertDialogPrimitive from '@radix-ui/react-alert-dialog';
 import * as React from 'react';
 
-import { buttonVariants } from '@/core-ui/components/ui/button';
-import { cn } from '@/core-ui/lib/utils';
+import { buttonVariants } from '@/components/ui/button';
+import { cn } from '@/lib/utils';
 
 const AlertDialogRoot = AlertDialogPrimitive.Root;
